rename auth package to mcpcontext
